moved BedType out of med model
