+++ care/facility/models/daily_round.py
@@ -17,11 +17,7 @@ from care.facility.models.json_schema.daily_round import (
     OUTPUT,
     PRESSURE_SORE,
 )
-from care.facility.models.patient_base import (
-    ADMIT_CHOICES,
-    CURRENT_HEALTH_CHOICES,
-    SYMPTOM_CHOICES,
-)
+from care.facility.models.patient_base import CURRENT_HEALTH_CHOICES, SYMPTOM_CHOICES
 from care.facility.models.patient_consultation import PatientConsultation
 from care.users.models import User
 from care.utils.models.validators import JSONFieldSchemaValidator
