+++ care/facility/models/patient_base.py
@@ -99,7 +99,19 @@ BLOOD_GROUP_CHOICES = [
 ]
 SuggestionChoices = SimpleNamespace(HI="HI", A="A", R="R", OP="OP", DC="DC")
 
+class BedType(enum.Enum):
+    ISOLATION = 1
+    ICU = 2
+    ICU_WITH_NON_INVASIVE_VENTILATOR = 3
+    ICU_WITH_OXYGEN_SUPPORT = 4
+    ICU_WITH_INVASIVE_VENTILATOR = 5
+    BED_WITH_OXYGEN_SUPPORT = 6
+    REGULAR = 7
+
+BedTypeChoices = [(e.value, e.name) for e in BedType]
+
 REVERSE_BLOOD_GROUP_CHOICES = reverse_choices(BLOOD_GROUP_CHOICES)
 REVERSE_DISEASE_STATUS_CHOICES = reverse_choices(DISEASE_STATUS_CHOICES)
 REVERSE_SYMPTOM_CATEGORY_CHOICES = reverse_choices(CATEGORY_CHOICES)
 REVERSE_ADMIT_CHOICES = reverse_choices(ADMIT_CHOICES)
+REVERSE_BED_TYPE_CHOICES = reverse_choices(BedTypeChoices)
